Add saveOnly entry point
