+++ package.json
@@ -3,7 +3,7 @@
   "main": "lib/restoreCache.js",
   "license": "MIT",
   "scripts": {
-    "build": "tsc && ncc build -o dist/restore src/restore.ts && ncc build -o dist/save src/save.ts",
+    "build": "tsc && ncc build -o dist/restore src/restore.ts && ncc build -o dist/save src/save.ts && ncc build -o dist/saveOnly src/saveOnly.ts",
     "test": "tsc --noEmit && jest --coverage"
   },
   "dependencies": {
