+++ save/action.yml
@@ -47,7 +47,7 @@ inputs:
   #   default: -n
 runs:
   using: node20
-  main: "../dist/save/index.js"
+  main: "../dist/saveOnly/index.js"
 branding:
   icon: "archive"
   color: "gray-dark"
