+++ src/utils.ts
@@ -3,6 +3,9 @@ import * as utils from "@actions/cache/lib/internal/cacheUtils";
 import * as core from "@actions/core";
 import * as minio from "minio";
 import { State } from "./state";
+import path from "path";
+import {createTar, listTar} from "@actions/cache/lib/internal/tar";
+import * as cache from "@actions/cache";
 
 export function isGhes(): boolean {
   const ghUrl = new URL(
@@ -180,4 +183,66 @@ export function isExactKeyMatch(): boolean {
     `isExactKeyMatch: matchedKey=${matchedKey} inputKey=${inputKey}, result=${result}`
   );
   return result;
+}
+
+export async function saveCache(standalone: boolean) {
+  try {
+    if (!standalone && isExactKeyMatch()) {
+      core.info("Cache was exact key match, not saving");
+      return;
+    }
+
+    const bucket = core.getInput("bucket", { required: true });
+    // Inputs are re-evaluted before the post action, so we want the original key
+    const key = standalone ? core.getInput("key", { required: true }) : core.getState(State.PrimaryKey);
+    const useFallback = getInputAsBoolean("use-fallback");
+    const paths = getInputAsArray("path");
+
+    try {
+      const mc = newMinio({
+        // Inputs are re-evaluted before the post action, so we want the original keys & tokens
+        accessKey: standalone ? getInput("accessKey", "AWS_ACCESS_KEY_ID") : core.getState(State.AccessKey),
+        secretKey: standalone ? getInput("secretKey", "AWS_SECRET_ACCESS_KEY") : core.getState(State.SecretKey),
+        sessionToken: standalone ? getInput("sessionToken", "AWS_SESSION_TOKEN") : core.getState(State.SessionToken),
+        region: standalone ? getInput("region", "AWS_REGION") : core.getState(State.Region),
+      });
+
+      const compressionMethod = await utils.getCompressionMethod();
+      const cachePaths = await utils.resolvePaths(paths);
+      core.debug("Cache Paths:");
+      core.debug(`${JSON.stringify(cachePaths)}`);
+
+      const archiveFolder = await utils.createTempDirectory();
+      const cacheFileName = utils.getCacheFileName(compressionMethod);
+      const archivePath = path.join(archiveFolder, cacheFileName);
+
+      core.debug(`Archive Path: ${archivePath}`);
+
+      await createTar(archiveFolder, cachePaths, compressionMethod);
+      if (core.isDebug()) {
+        await listTar(archivePath, compressionMethod);
+      }
+
+      const object = path.join(key, cacheFileName);
+
+      core.info(`Uploading tar to s3. Bucket: ${bucket}, Object: ${object}`);
+      await mc.fPutObject(bucket, object, archivePath, {});
+      core.info("Cache saved to s3 successfully");
+    } catch (e) {
+      core.info("Save s3 cache failed: " + e.message);
+      if (useFallback) {
+        if (isGhes()) {
+          core.warning("Cache fallback is not supported on Github Enterpise.");
+        } else {
+          core.info("Saving cache using fallback");
+          await cache.saveCache(paths, key);
+          core.info("Save cache using fallback successfully");
+        }
+      } else {
+        core.debug("skipped fallback cache");
+      }
+    }
+  } catch (e) {
+    core.info("warning: " + e.message);
+  }
 }
